+++ Imperium/src/Netcode/ImpNetworking.cs
@@ -1,5 +1,6 @@
 #region
 
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
@@ -138,13 +139,14 @@ public class ImpNetworking
     [ImpAttributes.LocalMethod]
     private void OnAuthenticateResponse()
     {
-        // actually authenticated client ID
-        Imperium.IO.Send(
-            "Imperium access was granted!",
-            type: NotificationType.AccessControl
-        );
-        Imperium.IO.LogInfo("[NET] Imperium access was granted!");
-        Imperium.Launch();
+        Imperium.IO.Send("Imperium access was granted!", type: NotificationType.AccessControl);
+        Imperium.IO.LogInfo("[NET] Imperium access was granted! Launching Imperium...");
+
+        if (!ImpUtils.RunSafeInProd(Imperium.Launch, out var exception))
+        {
+            Imperium.DisableImperium();
+            Imperium.IO.LogError($"[ERR] Imperium startup failed: {exception.Message}");
+        }
 
         // Request network values update from server if client is not host
         if (!NetworkManager.Singleton.IsHost) clientRequestValues.DispatchToServer();
@@ -155,11 +157,7 @@ public class ImpNetworking
     {
         if (NetworkManager.Singleton.IsHost) return;
 
-        Imperium.IO.Send(
-            "Imperium access was revoked!",
-            type: NotificationType.AccessControl,
-            isWarning: true
-        );
+        Imperium.IO.Send("Imperium access was revoked!", type: NotificationType.AccessControl, isWarning: true);
         Imperium.IO.LogInfo("[NET] Imperium access was revoked!");
         Imperium.DisableImperium();
     }
@@ -169,18 +167,19 @@ public class ImpNetworking
     {
         if (NetworkManager.Singleton.IsHost) return;
 
-        Imperium.IO.Send(
-            "Imperium access was granted!",
-            type: NotificationType.AccessControl
-        );
-        Imperium.IO.LogInfo("[NET] Imperium access was granted!");
+        Imperium.IO.Send("Imperium access was granted!", type: NotificationType.AccessControl);
+        Imperium.IO.LogInfo("[NET] Imperium access was granted! Launching Imperium...");
         if (Imperium.WasImperiumAccessGranted)
         {
             Imperium.EnableImperium();
         }
         else
         {
-            Imperium.Launch();
+            if (!ImpUtils.RunSafeInProd(Imperium.Launch, out var exception))
+            {
+                Imperium.DisableImperium();
+                Imperium.IO.LogError($"[ERR] Imperium startup failed: {exception.Message}");
+            }
 
             // Request network values update from server if client is not host
             if (!NetworkManager.Singleton.IsHost) clientRequestValues.DispatchToServer();